NOTE! Commit Description
Added findStudent method in Application.
Row 197(Application), returns StudentName when calling (s.getSSN). And returns StudentSSN when (s.getStudentName). Reason for this is that input from Application, does not allign with SQL Columns. A ssn,name,adress input in add Student is shown as adress, ssn, name in the database??